mqtt-demo-server/requirements.txt
# Python dependencies for MQTT Demo Server
paho-mqtt>=1.6.1